move oauth authenticators to their own namespace
