+++ index.d.ts
@@ -7,8 +7,8 @@
 /// <reference path="lib/Http/Client.d.ts" />
 /// <reference path="lib/Http/ClientInterface.d.ts" />
 /// <reference path="lib/Requester/Decorator/Authentication/HttpBasicAuthenticator.d.ts" />
-/// <reference path="lib/Requester/Decorator/Authentication/OAuthClientTokenAuthenticator.d.ts" />
-/// <reference path="lib/Requester/Decorator/Authentication/OAuthTokenPasswordAuthenticator.d.ts" />
+/// <reference path="lib/Requester/Decorator/Authentication/OAuth/ClientTokenAuthenticator.d.ts" />
+/// <reference path="lib/Requester/Decorator/Authentication/OAuth/TokenPasswordAuthenticator.d.ts" />
 /// <reference path="lib/Requester/Decorator/BodyConverterDecorator.d.ts" />
 /// <reference path="lib/Requester/Decorator/DecoratorInterface.d.ts" />
 /// <reference path="lib/Requester/Decorator/UrlDecorator.d.ts" />
