+++ lib/Api/Client.js
@@ -3,7 +3,7 @@ import '@jymfony/util/lib/String/sprintf';
 
 import BaseClient from '../Http/Client';
 import BodyConverterDecorator from '../Requester/Decorator/BodyConverterDecorator';
-import OAuthTokenPasswordAuthenticator from '../Requester/Decorator/Authentication/OAuthTokenPasswordAuthenticator';
+import TokenPasswordAuthenticator from '../Requester/Decorator/Authentication/OAuth/TokenPasswordAuthenticator';
 import UrlDecorator from '../Requester/Decorator/UrlDecorator';
 import VersionSetterDecorator from '../Requester/Decorator/VersionSetterDecorator';
 
@@ -29,7 +29,7 @@ export default class Client extends BaseClient {
         }
 
         const decorators = [];
-        const authenticator = new OAuthTokenPasswordAuthenticator(requester, tokenStorage, {
+        const authenticator = new TokenPasswordAuthenticator(requester, tokenStorage, {
             token_endpoint: (new URL('/token', config.base_url)).toString(),
             client_id: config.client_id,
             client_secret: config.client_secret,
