+++ lib/Http/Client.js
@@ -8,7 +8,7 @@ import HttpException from '../Exception/HttpException';
 import NotFoundHttpException from '../Exception/NotFoundHttpException';
 
 /**
- * @memberOf Fazland.Atlante.Api
+ * @memberOf Fazland.Atlante.Http
  * @implements ClientInterface
  */
 export default class Client {
@@ -16,7 +16,7 @@ export default class Client {
      * Constructor.
      *
      * @param {RequesterInterface} requester Object which performs http requests.
-     * @param {DecoratorInterface[]} requestDecorators Decorators to apply to the request.
+     * @param {DecoratorInterface[]} [requestDecorators = []] Decorators to apply to the request.
      */
     constructor(requester, requestDecorators = []) {
         /**
