+++ lib/Requester/Decorator/Authentication/OAuth/ClientTokenAuthenticator.js
@@ -1,7 +1,10 @@
-import Headers from '../../Headers';
-import NoTokenAvailableException from '../../../Exception/NoTokenAvailableException';
+import Headers from '../../../Headers';
+import NoTokenAvailableException from '../../../../Exception/NoTokenAvailableException';
 
-export default class OAuthClientTokenAuthenticator {
+/**
+ * @memberOf Fazland.Atlante.Requester.Decorator.Authentication.OAuth
+ */
+export default class ClientTokenAuthenticator {
     /**
      * Constructor.
      *
@@ -9,10 +12,9 @@ export default class OAuthClientTokenAuthenticator {
      * @param {Fazland.Atlante.Storage.StorageInterface} tokenStorage
      * @param {string} token_endpoint
      * @param {string} client_id
-     * @param {null|string} [client_secret]
-     * @param {string} [client_token_key]
-     * @param {string} [data_encoding] could be "json" or "form". If "json" encodes token request data with JSON.stringify,
-     *                  if "form" data will be encoded as form data.
+     * @param {null|string} [client_secret = null]
+     * @param {string} [client_token_key = 'fazland_atlante_client_token']
+     * @param {string} [data_encoding = 'json'] could be "json" or "form". If "json" encodes token request data with JSON.stringify, if "form" data will be encoded as form data.
      */
     constructor(
         requester,
