+++ lib/Requester/Decorator/Authentication/OAuth/TokenPasswordAuthenticator.js
@@ -1,7 +1,7 @@
-import NoTokenAvailableException from '../../../Exception/NoTokenAvailableException';
-import OAuthClientTokenAuthenticator from './OAuthClientTokenAuthenticator';
+import NoTokenAvailableException from '../../../../Exception/NoTokenAvailableException';
+import ClientTokenAuthenticator from './ClientTokenAuthenticator';
 
-export default class OAuthTokenPasswordAuthenticator extends OAuthClientTokenAuthenticator {
+export default class TokenPasswordAuthenticator extends ClientTokenAuthenticator {
     /**
      * Constructor.
      *
@@ -12,8 +12,8 @@ export default class OAuthTokenPasswordAuthenticator extends OAuthClientTokenAut
      * @param {null|string} [client_secret]
      * @param {string} [client_token_key]
      * @param {string} [data_encoding]
-     * @param {string} [access_token_key]
-     * @param {string} [refresh_token_key]
+     * @param {string} [access_token_key = 'access_token']
+     * @param {string} [refresh_token_key = 'refresh_token']
      */
     constructor(
         requester,
