+++ lib/Requester/Header/BaseAccept.js
@@ -109,7 +109,7 @@ export default class BaseAccept {
 
     /**
      * @param {string} key
-     * @param {*} default_
+     * @param {*} [default_ = null]
      *
      * @returns {string|*}
      */
@@ -118,6 +118,8 @@ export default class BaseAccept {
     }
 
     /**
+     * Checks if a parameter is set.
+     *
      * @param {string} key
      *
      * @returns {boolean}
