+++ lib/api.js
@@ -9,8 +9,8 @@ import NoTokenAvailableException from './Exception/NoTokenAvailableException';
 import HttpClient from './Http/Client';
 
 import HttpBasicAuthenticator from './Requester/Decorator/Authentication/HttpBasicAuthenticator';
-import OAuthClientTokenAuthenticator from './Requester/Decorator/Authentication/OAuthClientTokenAuthenticator';
-import OAuthTokenPasswordAuthenticator from './Requester/Decorator/Authentication/OAuthTokenPasswordAuthenticator';
+import ClientTokenAuthenticator from './Requester/Decorator/Authentication/OAuth/ClientTokenAuthenticator';
+import TokenPasswordAuthenticator from './Requester/Decorator/Authentication/OAuth/TokenPasswordAuthenticator';
 
 import BodyConverterDecorator from './Requester/Decorator/BodyConverterDecorator';
 import UrlDecorator from './Requester/Decorator/UrlDecorator';
@@ -48,9 +48,11 @@ export default {
     Requester: {
         Decorator: {
             Authentication: {
+                OAuth: {
+                    ClientTokenAuthenticator,
+                    TokenPasswordAuthenticator,
+                },
                 HttpBasicAuthenticator,
-                OAuthClientTokenAuthenticator,
-                OAuthTokenPasswordAuthenticator,
             },
             BodyConverterDecorator,
             UrlDecorator,
